refactor(services): keep remote model requests on stream services
